tom/re-structure_project: refactor the play module folder structure
