--- managementapp-play/src/main/resources/application.conf
@@ -1,1 +1,0 @@
-# https://www.playframework.com/documentation/latest/Configuration
