+++ play/src/main/resources/application.conf
@@ -1,0 +1,3 @@
+# https://www.playframework.com/documentation/latest/Configuration
+
+# MySQL config
