+++ play/src/main/scala/com/managementapp/controllers/root/ManagementAppController.scala
@@ -1,8 +1,8 @@
-package com.managementapp.controllers
+package com.managementapp.controllers.root
 
 import com.google.inject.{Inject, Singleton}
-import play.api.mvc._
 import play.Configuration
+import play.api.mvc._
 
 @Singleton
 class ManagementAppController @Inject() (configuration: Configuration) extends Controller {
