+++ quiz-core/src/main/scala/com/managementapp/controllers/quiz/Quiz.scala
@@ -1,4 +1,4 @@
-package com.managementapp.quiz
+package com.managementapp.controllers.quiz
 
 /**
   * Created by omzmarlon on 2017-03-29.
